requirements.txt
fastapi
uvicorn
sqlmodel
azure-identity
python-jose[cryptography]
passlib[bcrypt]
python-multipart